+++ src/libraries/System.Net.Http/src/System/Net/WebTransport/WebTransportException.cs
@@ -3,20 +3,28 @@
 
 namespace System.Net.WebTransport;
 
+// TODO: redo exception hierarchy similar to QUIC
 public class WebTransportException : Exception
 {
     public WebTransportException(string message) : base(message) { }
+
     public WebTransportException(string message, Exception innerException) : base(message, innerException) { }
 }
 
 public sealed class WebTransportStreamClosedException : WebTransportException
 {
-    public int ApplicationErrorCode { get; }
-    public WebTransportStreamClosedException(string message, int applicationErrorCode) : base(message)
+    /// <summary>
+    /// Error code provided by the peer when closing the stream.
+    /// The value is in the range [0, 2^32).
+    /// </summary>
+    public long ApplicationErrorCode { get; }
+
+    public WebTransportStreamClosedException(string message, long applicationErrorCode) : base(message)
     {
         ApplicationErrorCode = applicationErrorCode;
     }
-    public WebTransportStreamClosedException(string message, int applicationErrorCode, Exception innerException) : base(message, innerException)
+
+    public WebTransportStreamClosedException(string message, long applicationErrorCode, Exception innerException) : base(message, innerException)
     {
         ApplicationErrorCode = applicationErrorCode;
     }
@@ -26,11 +34,13 @@ public sealed class WebTransportSessionClosedException : WebTransportException
 {
     public int ApplicationErrorCode { get; }
     public string ApplicationErrorMessage { get; }
+
     public WebTransportSessionClosedException(string message, int applicationErrorCode, string applicationErrorMessage) : base(message)
     {
         ApplicationErrorCode = applicationErrorCode;
         ApplicationErrorMessage = applicationErrorMessage;
     }
+
     public WebTransportSessionClosedException(string message, int applicationErrorCode, string applicationErrorMessage, Exception innerException) : base(message, innerException)
     {
         ApplicationErrorCode = applicationErrorCode;
